Fixes #128
